+++ src/django/api/reports/average_list_count_per_contributor.sql
@@ -12,6 +12,7 @@ FROM (
   WHERE to_char(l.created_at, 'YYYY-MM') != to_char(now(), 'YYYY-MM')
   AND s.is_public = 't'
   AND s.is_active = 't'
+  AND s.create = 't'
   GROUP BY to_char(l.created_at, 'YYYY-MM'), c.id
 ) s
 GROUP BY month