+++ src/django/api/reports/contributor_first_appearance.sql
@@ -1,16 +1,17 @@
 SELECT
   month,
   name
-from (
+FROM (
   SELECT DISTINCT
-    min(to_char(l.created_at, 'YYYY-MM')) AS month,
+    min(to_char(s.created_at, 'YYYY-MM')) AS month,
     c.id
-  FROM api_facilitylist l
-  JOIN api_source s ON s.facility_list_id = l.id
+  FROM api_source s
   JOIN api_contributor c ON s.contributor_id = c.id
   JOIN api_user u ON u.id = c.admin_id
-  WHERE to_char(l.created_at, 'YYYY-MM') != to_char(now(), 'YYYY-MM')
+  WHERE to_char(s.created_at, 'YYYY-MM') != to_char(now(), 'YYYY-MM')
+  AND s.create = 't'
   AND NOT u.email LIKE '%openapparel.org%'
+  AND NOT u.email LIKE '%azavea.com%'
   GROUP BY c.id
 ) s
 JOIN api_contributor c ON c.id = s.id